requirements.txt
easydict>=1.9
medpy>=0.4.0
opencv-python>=4.5.0
pytorch-lightning>=2.0.0
timm>=0.4.10
torchmetrics>=0.11.0
scipy>=1.7.0
tqdm>=4.40.0
tensorboardX>=2.2
packaging
wandb>=0.15.0
timm==0.4.12
pytest 
chardet 
yacs 
termcolor
submitit 
tensorboardX
fvcore
seaborn
einops
